+++ SW2URDF/URDFExport/ExportPropertyManagerExtension.cs
@@ -636,14 +636,6 @@ namespace SW2URDF.URDFExport
 
             SetConfigTree(baseNode);
 
-            IPropertyManagerPageControl loadConfigurationControl = (IPropertyManagerPageControl)PMButtonLoad;
-
-            if (baseNode == null || !baseNode.RebuildLink().AreRequiredFieldsSatisfied())
-            {
-                loadConfigurationControl.Tip = 
-                    "Your configuration has not been fully exported. This feature may not work correctly";
-            }
-
             return true;
         }